added initial config for orders entity

+++ src/main/java/com/lucianna/mendonca/onlineorderformwebapp/repository/CustomerRepository.java
@@ -3,5 +3,5 @@ package com.lucianna.mendonca.onlineorderformwebapp.repository;
 import com.lucianna.mendonca.onlineorderformwebapp.model.CustomerModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CustomerRepository extends JpaRepository<CustomerModel, Integer> {
+public interface CustomerRepository extends JpaRepository<CustomerModel, Long> {
 }
