+++ src/main/java/com/lucianna/mendonca/onlineorderformwebapp/repository/OrderRepository.java
@@ -3,5 +3,5 @@ package com.lucianna.mendonca.onlineorderformwebapp.repository;
 import com.lucianna.mendonca.onlineorderformwebapp.model.OrderModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrderRepository extends JpaRepository<OrderModel, Integer> {
+public interface OrderRepository extends JpaRepository<OrderModel, Long> {
 }
